requirements.txt
Flask
gunicorn
biopython